requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
aiohttp==3.8.5; python_version >= '3.6'
aiosignal==1.3.1; python_version >= '3.7'
async-timeout==4.0.3; python_version < '3.11'
attrs==23.1.0; python_version >= '3.7'
beautifulsoup4==4.12.2
blinker==1.6.2; python_version >= '3.7'
cachetools==5.3.1
cdifflib==1.2.6
certifi==2023.7.22; python_version >= '3.6'
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
commonmark==0.9.1
dataclasses-json==0.5.14; python_version < '3.13' and python_version >= '3.7'
deprecated==1.2.14
dnspython==2.4.2; python_version >= '3.8' and python_version < '4.0'
filelock==3.12.2; python_version >= '3.7'
flask-cors==4.0.0
flask-limiter==3.3.1
flask==2.3.2
frozenlist==1.4.0; python_version >= '3.8'
fsspec==2023.6.0; python_version >= '3.8'
greenlet==2.0.2
huggingface-hub==0.16.4; python_version >= '3.7'
idna==3.4
importlib-resources==6.0.1; python_version >= '3.8'
itsdangerous==2.1.2
jinja2==3.1.2
joblib==1.3.2; python_version >= '3.7'
langchain==0.0.266; python_full_version >= '3.8.1' and python_version < '4.0'
langsmith==0.0.25; python_full_version >= '3.8.1' and python_version < '4.0'
levenshtein==0.21.1; python_version >= '3.6'
limits==3.5.0
llama-index==0.8.5.post2
loguru==0.7.0; python_version >= '3.5'
markdown-it-py==3.0.0; python_version >= '3.8'
markupsafe==2.1.3
marshmallow==3.20.1; python_version >= '3.8'
mdurl==0.1.2; python_version >= '3.7'
multidict==6.0.4; python_version >= '3.7'
mypy-extensions==1.0.0; python_version >= '3.5'
nest-asyncio==1.5.7; python_version >= '3.5'
nltk==3.8.1
numexpr==2.8.5; python_version >= '3.7'
numpy==1.25.2; python_version >= '3.10'
openai==0.27.8; python_full_version >= '3.7.1'
openapi-schema-pydantic==1.2.4; python_full_version >= '3.6.1'
ordered-set==4.1.0
packaging==23.1; python_version >= '3.7'
pandas==2.0.3
patsy==0.5.3
pinecone-client==2.2.2
psycopg2==2.9.7
pydantic==1.10.12; python_version >= '3.7'
pygments==2.16.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-levenshtein==0.21.1
pytz==2023.3
pyyaml==6.0.1; python_version >= '3.6'
rapidfuzz==3.2.0; python_version >= '3.7'
regex==2023.8.8; python_version >= '3.6'
requests==2.31.0
rich==13.5.2
safetensors==0.3.2
scipy==1.11.2
setuptools==68.1.2; python_version >= '3.8'
six==1.16.0
soupsieve==2.4.1; python_version >= '3.7'
sqlalchemy==2.0.20
statsmodels==0.14.0
tenacity==8.2.3; python_version >= '3.7'
tiktoken==0.4.0; python_version >= '3.8'
tokenizers==0.13.3
tqdm==4.66.1; python_version >= '3.7'
transformers==4.31.0
typing-extensions==4.7.1
typing-inspect==0.9.0
tzdata==2023.3; python_version >= '2'
urllib3==1.26.16
waitress==2.1.2
werkzeug==2.3.7
win32-setctime==1.1.0; sys_platform == 'win32'
wrapt==1.15.0
yarl==1.9.2; python_version >= '3.7'
